move pretty print helpers to utils
